Add types file
