+++ packages/connect-react/src/components/Errors.tsx
@@ -13,12 +13,8 @@ export type ErrorsProps<T extends ConfigurableProps, U extends ConfigurableProp>
 export function Errors<T extends ConfigurableProps, U extends ConfigurableProp>(props: ErrorsProps<T, U>) {
   const { field } = props;
   const {
-    errors, prop,
+    errors = {}, prop = {},
   } = field
-
-  if (!Object.keys(errors || {}).length) {
-    return null;
-  }
 
   if (!errors[prop.name]) {
     return null
