+++ packages/connect-react/src/types.ts
@@ -1,0 +1,11 @@
+export type SdkError = {
+    name: string;
+    message: string;
+}
+
+export type Observation = {
+    ts: number;
+    k: string;
+    h?: string;
+    err?: Error;
+}
