+++ types/src/index.ts
@@ -397,15 +397,3 @@ export function defineAction<
 (component: Action<Methods, ActionPropDefinitions>): Action<Methods, ActionPropDefinitions> {
   return component;
 }
-
-export type SdkError = {
-  name: string;
-  message: string;
-}
-
-export type Observation = {
-  ts: number;
-  k: string;
-  h?: string;
-  err?: Error;
-}
